Fix region span calculations
If one region provided to span is a subset of the other, then the
result should be the larger enclosing region.  Instead one of the
endpoints was from the smaller enclosed region.

This was being triggered by a strange evaluation order.  If we have
consecutive regions A, B, C, D, then it would compute the span of (A, D)
then the span of that with C, then the span of that with B.  Change it
to compute left-to-right:  A with B, result with C, result with D.