Split export.sml to constituent tools
Drop the noweb tool since it is built-in for all recent SML/NJ versions.
For all other tools, switch to Tools.registerStdShellCmdTool to do the
work.  I've copied most features except for removing files upon failure
(not clear whether this happens automatically, or how to do it if not).

The *-ext.cm and *-tool.cm files are intended to be loaded as plugins
whenever CM needs to handle a file with the given suffix or class,
respectively.  There should be no need to build a separate heap
containing these tools and the extra stuff that was at the end of
export.sml.