Allow building without tool installation
Move the ml-lex+ heap creation to tools/Makefile.  Delete the
installation of that heap; instead we use @SMLload to run it directly.

The CM tools are no longer portable since they now have relative paths
(referencing files in ../tools when run from the src directory), but we
don't care as long as we don't need to install them.

Observation:  The install script doesn't generate nor install a heap for
the toolkit sources themselves, so there is not much left for it to do.
Replace it with src/Makefile.  Also add the heap creation.