Heavy-handed workaround for EXP_BASICS structure sharing
If a signature contains both rigid and flexible types, in our case:

    signature EXP_BASICS = sig
      type exp
      type number = int
      ...
    end

And then we create other signatures with such a structure as a member
AND we impose sharing constraints between the various copies, such as:

    signature PATTERN = sig
      structure Exp : EXP_SET (* includes EXP_BASICS indirectly *)
      ...
    end

    signature PATTERNS = sig
      structure Exp : EXP_SET
      structure Pure : PATTERN
      structure Labelled : PATTERN
      ...
      sharing Exp = Pure.Exp = Labelled.Exp = ...
      ...
    end

Then it becomes difficult or impossible to specify the flexible member
after the fact, because it will complain that the rigid members cannot
be shared:

    functor Fun(structure Patterns : PATTERNS where type Exp.exp = ...) = ...

Since the exp type in EXP_BASICS is only ever defined as UnifiedExp.exp,
just drop the ability to keep that type flexible and define it as
UnifiedExp.exp.  If there is a better solution, please revert this
commit and fix it properly.