Repair/remove sharing in MainFun parameters
All of the 'where type' and 'where structure' clauses for the Code
members are incompatible with the sharing constraint at the end of the
parameters, and with other definitions in the signatures.

Fortunately all parameter structures are created by LinkEncoderFun which
sets the Code members of all of them to ULinked.Code.  So instead define
all Code types to match ULinked.Code.

Remove the extra Gencode argument from the MainFun calls.