Avoid TargetWord.toIntX when printing
When Int.precision >= TargetWord.wordSize (particularly on 64-bit hosts)
AND when TargetWord bit 31 is set, then this prints a negative number,
usually in SML syntax (e.g. ~1), except for ppcon' in emit-c.nw that
changes the '~' to '-'.

Previous behavior generated code such as

  unsigned MATCH_w_32_0;
  {
    MATCH_w_32_0 = getDword(MATCH_p);
    if ((MATCH_w_32_0 & 0xffffffff)
            /* op*rd*op3*rs1*i*simm13 at 0 */ == -2117607416) {

If integers are larger than 32 bits, then the condition would never be
true since bits 32+ on the right-hand side are sign-extended to 1, but
0 on the left-hand side.

Change most printing to be in decimal to match previous behavior.