Fix recursion bounds
Range lo and hi limits were previously (signed) integers and are now
(unsigned) words.  The recursion in addRange is adding case labels for
(hi-1) downto lo, and was relying on being able to set hi less than lo
to terminate the recursion.  This is no longer possible when lo is 0;
instead hi underflows to 0wxffffffff or similar and still compares
greater than or equal to lo.