Don't capture the newline character in CODELINE tokens
Strings contained in these lexer tokens always end with "\n".  They
eventually make it to the prettyprinter, and we are inconsistently
appending additional prettyprinter line breaks:

    RAW statements (inside match arms):  pp.nl appended via pplines
    Code outside of any match blocks:    no extra breaks inserted

Line breaks in the match arms therefore get doubled in the output.

To make it consistent, we could skip appending the additional line
breaks.  But the prettyprinter keeps track of line numbers for synch
support, and it uses its own BREAK and SYNCH tokens exclusively to
count lines.  Any "\n" in the input would throw off the count.  So
instead we need to replace all the "\n"s with prettyprinter breaks:
Don't capture "\n" while lexing, do append pp.nl to each line.