Fix missing braces on some match arms
From what I can tell, it's designed to add braces for the "opt-block"
arms, but not the "opt-block+" arm.  "opt-block+" is a BLOCK with no
definitions and a single statement that is not a RAW statement (i.e.
text from the matcher file that may actually hold multiple statements).

The problem is that every RAW is wrapped in STATEMENT' and therefore
occurrences of these are actually matching the "opt-block+" pattern.