Avoid extra indenting for single-arm IF statements
Seems to occur when an IF contains an INTCASE; I assume the INTCASE
makes the IF obsolete, and the IF isn't or can't be removed for some
reason.  But the net effect is that it would still emit prettyprinter
tokens for the IF, creating an empty line and extra indent for the
INTCASE it contains.

Simplify all of the IF code to better handle the "else" with no guard,
and delete the "opt-block" comments.