Insert a space when an optional break is not taken
This is usually the preferred behavior, so now it is the default for $o
in the format strings.  Use $O if you need to omit this space, similar
to how $N omits space compared to $n.

Remove several manual spaces around $o or pp.on in format strings/lists.
Whitespace before can almost always be deleted; whitespace after may
cause strange indenting on the following line when the break is taken.